working on getting comments controller working also seperated comments and posts into there own controllers
