+++ binturong-api/src/main/java/com/binturong/demo/repositorys/PostRepository.java
@@ -1,9 +1,11 @@
 package com.binturong.demo.repositorys;
 
 import com.binturong.demo.entities.Post;
+import com.binturong.demo.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post,Integer> {
+    Post findAllById(Integer postId);
 }
