+++ binturong-api/src/main/java/com/binturong/demo/services/CommentService.java
@@ -1,6 +1,7 @@
 package com.binturong.demo.services;
 
 import com.binturong.demo.entities.Comments;
+import com.binturong.demo.entities.Post;
 import com.binturong.demo.entities.User;
 
 import java.util.List;
@@ -9,5 +10,5 @@ public interface CommentService {
 
     public Comments saveComments(Comments comment);
 
-    public List<Comments> getPostComments(User user);
+    public List<Comments> getPostComments(Post post);
 }
